Addressing a few issues raised by @aduth
