+++ assets/stylesheets/_components.scss
@@ -9,6 +9,7 @@
 @import 'layout/style';
 
 @import 'auth/style';
+@import 'blocks/app-promo/style';
 @import 'blocks/author-compact-profile/style';
 @import 'blocks/author-selector/style';
 @import 'blocks/comment-button/style';
@@ -26,7 +27,6 @@
 @import 'blocks/reader-search-card/style';
 @import 'components/accordion/style';
 @import 'components/animate/style';
-@import 'components/app-promo/style';
 @import 'components/bulk-select/style';
 @import 'components/button/style';
 @import 'components/button-group/style';
