+++ client/blocks/app-promo/style.scss
