+++ client/reader/sidebar/test/index.jsx
@@ -7,14 +7,14 @@ import useFakeDom from 'test/helpers/use-fake-dom';
 import useMockery from 'test/helpers/use-mockery';
 
 describe( 'ReaderSidebar', ( ) => {
-	let ReaderSidebar
+	let ReaderSidebar;
 
 	let shouldRenderAppPromo;
 	const shouldRenderAppPromoDefaultProps = {
 		isDesktopPromoDisabled: false,
 		isUserLocaleEnglish: true,
 		isViewportMobile: false,
-		isUserOnChromeOS: false,
+		isUserOnChromeOs: false,
 		isDesktopPromoConfiguredToRun: true,
 		isUserDesktopAppUser: false
 	};
@@ -35,18 +35,17 @@ describe( 'ReaderSidebar', ( ) => {
 	} );
 
 	context( 'AppPromo', ( ) => {
-
 		it( 'should render the AppPromo when the shouldRenderAppPromo property is true', ( ) => {
-			const adjustedProperties = { ...readerSidebarDefaultProps, shouldRenderAppPromo: true }
+			const adjustedProperties = { ...readerSidebarDefaultProps, shouldRenderAppPromo: true };
 			const wrapper = shallow( <ReaderSidebar { ...adjustedProperties } /> );
-			expect( wrapper.find( '.reader-sidebar__AppPromo' ) ).to.have.lengthOf( 1 );
-		});
+			expect( wrapper.find( '.sidebar__app-promo' ) ).to.have.lengthOf( 1 );
+		} );
 
 		it( 'should not render the AppPromo when the shouldRenderAppPromo property is false', ( ) => {
-			const adjustedProperties = { ...readerSidebarDefaultProps, shouldRenderAppPromo: false }
+			const adjustedProperties = { ...readerSidebarDefaultProps, shouldRenderAppPromo: false };
 			const wrapper = shallow( <ReaderSidebar { ...adjustedProperties } /> );
-			expect( wrapper.find( '.reader-sidebar__AppPromo' ) ).to.have.lengthOf( 0 );
-		});
+			expect( wrapper.find( '.sidebar__app-promo' ) ).to.have.lengthOf( 0 );
+		} );
 
 		context( 'shouldRenderAppPromo', ( ) => {
 			it( 'should not render if desktop promo is disabled', ( ) => {
@@ -62,7 +61,7 @@ describe( 'ReaderSidebar', ( ) => {
 			} );
 
 			it( 'should not render if it\'s ChromeOS', ( ) => {
-				expect( shouldRenderAppPromo( { ...shouldRenderAppPromoDefaultProps, isUserOnChromeOS: true } ) ).to.be.false;
+				expect( shouldRenderAppPromo( { ...shouldRenderAppPromoDefaultProps, isUserOnChromeOs: true } ) ).to.be.false;
 			} );
 
 			it( 'should not render if desktop promo isn\'t configured to run', ( ) => {
@@ -76,6 +75,6 @@ describe( 'ReaderSidebar', ( ) => {
 			it( 'should render if desktop promo wasn\'t disabled by the user, the locale is english, the viewport isn\'t mobile, it\'s not ChromeOS, the desktop promo is configured to run, and the user isn\'t a desktop app user', ( ) => {
 				expect( shouldRenderAppPromo( shouldRenderAppPromoDefaultProps ) ).to.be.true;
 			} );
-		});
+		} );
 	} );
 } );
